requirements.txt
numpy==2.2.5
pandas==2.2.3
matplotlib==3.10.3
oandapyV20==0.7.2
python-dotenv==1.1.0
tqdm==4.67.1
